switch extended arrays
