+++ tests/unit/model/arrays/test_array.py
@@ -11,10 +11,10 @@ from numpy.typing import NDArray
 
 from power_grid_model_ds._core import fancypy as fp
 from power_grid_model_ds._core.model.arrays.base.array import FancyArray
-from power_grid_model_ds._core.model.arrays.pgm_arrays import TransformerArray
+from power_grid_model_ds._core.model.arrays.pgm_arrays import LineArray, TransformerArray
 from power_grid_model_ds._core.model.constants import EMPTY_ID, empty
 from tests.conftest import FancyTestArray
-from tests.fixtures.arrays import ExtendedFancyTestArray, FancyTestArray3
+from tests.fixtures.arrays import ExtendedLineArray, FancyTestArray3
 
 # pylint: disable=missing-function-docstring
 
@@ -292,7 +292,7 @@ def test_overflow_value():
 
 
 def test_from_extended_array():
-    extended_array = ExtendedFancyTestArray.zeros(2)
-    array = FancyTestArray.from_extended(extended_array)
-    assert not isinstance(array, ExtendedFancyTestArray)
+    extended_array = ExtendedLineArray.zeros(2)
+    array = LineArray.from_extended(extended_array)
+    assert not isinstance(array, ExtendedLineArray)
     assert_array_equal(array.data, extended_array[array.columns])
